[Refactor] - Move LatLng-related code to GoogleMapUtil.kt as well
